Refactor modal handling for ticket operations
Added error handling with try functions whenever sending a modal. Why? There is a limit of in how much time you much return a response for button/option click to discord that is 3 seconds. Because of that, in high pings, bot used to give error saying "Unknown Interaction". This will fix that issue.